Move cfg.rs to store.rs
